# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ConsensusAI - Autonomous AI Governance and Decision Platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Override urgency classification
 * @nullable
 */
export type TaskInputUrgencyLevel = typeof TaskInputUrgencyLevel[keyof typeof TaskInputUrgencyLevel] | null;


export const TaskInputUrgencyLevel = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface TaskInput {
  /**
     * Operational command text
     * @minLength 1
     */
  command: string;
  /** Environmental/telemetry context JSON string */
  context: string;
  /**
     * Override urgency classification
     * @nullable
     */
  urgencyLevel?: TaskInputUrgencyLevel;
}

export interface OverrideInput {
  /** Human override action */
  action: string;
  /** Reason for override */
  rationale: string;
}

export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];


export const TaskStatus = {
  pending: 'pending',
  fast_path: 'fast_path',
  debating: 'debating',
  completed: 'completed',
  failed: 'failed',
  overridden: 'overridden',
} as const;

export type TaskRoutingPath = typeof TaskRoutingPath[keyof typeof TaskRoutingPath];


export const TaskRoutingPath = {
  fast_path: 'fast_path',
  cognitive_debate: 'cognitive_debate',
} as const;

export interface Task {
  id: number;
  command: string;
  context: string;
  status: TaskStatus;
  routingPath: TaskRoutingPath;
  urgencyScore: number;
  /** @nullable */
  finalAction: string | null;
  /** @nullable */
  consensusScore: number | null;
  /** @nullable */
  overrideAction?: string | null;
  /** @nullable */
  overrideRationale?: string | null;
  /** @nullable */
  executionTimeMs?: number | null;
  createdAt: string;
  updatedAt: string;
}

export type DeliberationAgentRole = typeof DeliberationAgentRole[keyof typeof DeliberationAgentRole];


export const DeliberationAgentRole = {
  planner: 'planner',
  risk: 'risk',
  resource: 'resource',
  ethics: 'ethics',
} as const;

export interface Deliberation {
  id: number;
  taskId: number;
  agentName: string;
  agentRole: DeliberationAgentRole;
  proposal: string;
  /** @nullable */
  reasoning?: string | null;
  priorityWeight: number;
  vetoed: boolean;
  /** @nullable */
  vetoReason?: string | null;
  confidenceScore?: number;
  createdAt: string;
}

export interface Capsule {
  id: number;
  taskId: number;
  blockHash: string;
  previousHash: string;
  blockIndex: number;
  /** @nullable */
  payload?: string | null;
  verified: boolean;
  createdAt: string;
}

export interface AuditExport {
  context: string;
  taskId: number;
  command: string;
  finalAction: string;
  capsule: Capsule;
  deliberations: Deliberation[];
  chainValid: boolean;
}

export interface DashboardStats {
  totalTasks: number;
  fastPathCount: number;
  cognitiveDebateCount: number;
  completedCount: number;
  failedCount: number;
  overriddenCount: number;
  avgConsensusScore: number;
  avgExecutionTimeMs: number;
  chainIntegrityRate: number;
}

export interface ActivityItem {
  id: number;
  command: string;
  status: string;
  routingPath: string;
  urgencyScore: number;
  /** @nullable */
  finalAction?: string | null;
  createdAt: string;
}