# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ConsensusAI - Autonomous AI Governance and Decision Platform
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityItem {
  id: number;
  command: string;
  status: string;
  routingPath: string;
  urgencyScore: number;
  /** @nullable */
  finalAction?: string | null;
  createdAt: string;
}