# lib/api-zod/src/generated/types/auditExport.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ConsensusAI - Autonomous AI Governance and Decision Platform
 * OpenAPI spec version: 0.1.0
 */
import type { Capsule } from './capsule';
import type { Deliberation } from './deliberation';

export interface AuditExport {
  context: string;
  taskId: number;
  command: string;
  finalAction: string;
  capsule: Capsule;
  deliberations: Deliberation[];
  chainValid: boolean;
}