# lib/api-zod/src/generated/types/capsule.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ConsensusAI - Autonomous AI Governance and Decision Platform
 * OpenAPI spec version: 0.1.0
 */

export interface Capsule {
  id: number;
  taskId: number;
  blockHash: string;
  previousHash: string;
  blockIndex: number;
  /** @nullable */
  payload?: string | null;
  verified: boolean;
  createdAt: string;
}