# lib/api-zod/src/generated/types/deliberation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ConsensusAI - Autonomous AI Governance and Decision Platform
 * OpenAPI spec version: 0.1.0
 */
import type { DeliberationAgentRole } from './deliberationAgentRole';

export interface Deliberation {
  id: number;
  taskId: number;
  agentName: string;
  agentRole: DeliberationAgentRole;
  proposal: string;
  /** @nullable */
  reasoning?: string | null;
  priorityWeight: number;
  vetoed: boolean;
  /** @nullable */
  vetoReason?: string | null;
  confidenceScore?: number;
  createdAt: string;
}