# lib/api-zod/src/generated/types/overrideInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ConsensusAI - Autonomous AI Governance and Decision Platform
 * OpenAPI spec version: 0.1.0
 */

export interface OverrideInput {
  /** Human override action */
  action: string;
  /** Reason for override */
  rationale: string;
}