# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ConsensusAI - Autonomous AI Governance and Decision Platform
 * OpenAPI spec version: 0.1.0
 */
import type { TaskRoutingPath } from './taskRoutingPath';
import type { TaskStatus } from './taskStatus';

export interface Task {
  id: number;
  command: string;
  context: string;
  status: TaskStatus;
  routingPath: TaskRoutingPath;
  urgencyScore: number;
  /** @nullable */
  finalAction: string | null;
  /** @nullable */
  consensusScore: number | null;
  /** @nullable */
  overrideAction?: string | null;
  /** @nullable */
  overrideRationale?: string | null;
  /** @nullable */
  executionTimeMs?: number | null;
  createdAt: string;
  updatedAt: string;
}