# lib/api-zod/src/generated/types/taskInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ConsensusAI - Autonomous AI Governance and Decision Platform
 * OpenAPI spec version: 0.1.0
 */
import type { TaskInputUrgencyLevel } from './taskInputUrgencyLevel';

export interface TaskInput {
  /**
     * Operational command text
     * @minLength 1
     */
  command: string;
  /** Environmental/telemetry context JSON string */
  context: string;
  /**
     * Override urgency classification
     * @nullable
     */
  urgencyLevel?: TaskInputUrgencyLevel;
}